+++ tests/middleware/test_catch_error.py
@@ -1,79 +1,75 @@
 import time
 
 import remoulade
-from remoulade.middleware.catch_error import CatchError
 
 
-def test_cleanup_actor(stub_broker, stub_worker):
-    cleanup_count = 0
-
-    @remoulade.actor
-    def cleanup(actor_name, exception_name, args, kwargs):
-        nonlocal cleanup_count
-        cleanup_count += 1
+def test_on_failure(stub_broker, stub_worker):
+    on_failure_count = 0
 
     @remoulade.actor
-    def fail_actor():
-        raise Exception()
-
-    remoulade.declare_actors([cleanup, fail_actor])
-
-    stub_broker.add_middleware(CatchError())
-    fail_actor.send_with_options(cleanup_actor=cleanup)
-
-    stub_broker.join(fail_actor.queue_name)
-    stub_broker.join(cleanup.queue_name)
-    stub_worker.join()
-
-    # The cleanup should have run
-    assert cleanup_count == 1
-
-
-def test_cleanup_runs_only_once(stub_broker, stub_worker):
-    cleanup_count = 0
-
-    @remoulade.actor
-    def cleanup(actor_name, exception_name, args, kwargs):
-        nonlocal cleanup_count
-        cleanup_count += 1
+    def on_failure(actor_name, exception_name, args, kwargs):
+        nonlocal on_failure_count
+        on_failure_count += 1
 
     @remoulade.actor
     def fail_actor():
         raise Exception()
 
-    remoulade.declare_actors([cleanup, fail_actor])
+    remoulade.declare_actors([on_failure, fail_actor])
 
-    stub_broker.add_middleware(CatchError())
-    fail_actor.send_with_options(cleanup_actor=cleanup, max_retries=3, min_backoff=1, retry_strategy="constant")
+    fail_actor.send_with_options(on_failure=on_failure)
 
     stub_broker.join(fail_actor.queue_name)
-    stub_broker.join(cleanup.queue_name)
+    stub_broker.join(on_failure.queue_name)
     stub_worker.join()
 
-    # The cleanup should have run only once
-    assert cleanup_count == 1
+    # The on_failure should have run
+    assert on_failure_count == 1
+
+
+def test_on_failure_runs_only_once(stub_broker, stub_worker):
+    on_failure_count = 0
+
+    @remoulade.actor
+    def on_failure(actor_name, exception_name, args, kwargs):
+        nonlocal on_failure_count
+        on_failure_count += 1
+
+    @remoulade.actor
+    def fail_actor():
+        raise Exception()
+
+    remoulade.declare_actors([on_failure, fail_actor])
+
+    fail_actor.send_with_options(on_failure=on_failure, max_retries=3, min_backoff=1, retry_strategy="constant")
+
+    stub_broker.join(fail_actor.queue_name)
+    stub_broker.join(on_failure.queue_name)
+    stub_worker.join()
+
+    # The on_failure should have run only once
+    assert on_failure_count == 1
 
 
 def test_clean_runs_on_timeout(stub_broker, stub_worker):
-    cleanup_count = 0
+    on_failure_count = 0
 
     @remoulade.actor
-    def cleanup(actor_name, exception_name, args, kwargs):
-        nonlocal cleanup_count
-        cleanup_count += 1
+    def on_failure(actor_name, exception_name, args, kwargs):
+        nonlocal on_failure_count
+        on_failure_count += 1
 
     @remoulade.actor
     def do_work():
         time.sleep(1)
 
-    remoulade.declare_actors([cleanup, do_work])
+    remoulade.declare_actors([on_failure, do_work])
 
-    stub_broker.add_middleware(CatchError())
-    do_work.send_with_options(cleanup_actor=cleanup, time_limit=1)
+    do_work.send_with_options(on_failure=on_failure, time_limit=1)
 
     stub_broker.join(do_work.queue_name)
-    stub_broker.join(cleanup.queue_name)
+    stub_broker.join(on_failure.queue_name)
     stub_worker.join()
 
-    # The cleanup should have run
-    assert cleanup_count == 1
+    # The on_failure should have run
+    assert on_failure_count == 1